Reorganize coo.py
