+++ sparse/__init__.py
@@ -1,10 +1,5 @@
-from .coo import COO, elemwise, tensordot, concatenate, stack, dot, triu, tril, where, \
-    nansum, nanprod, nanmin, nanmax
+from .coo import *
 from .dok import DOK
 from .sparse_array import SparseArray
 from .utils import random
 from ._version import __version__
-
-__all__ = ["SparseArray", "COO", "DOK",
-           "tensordot", "concatenate", "stack", "dot", "triu", "tril", "random", "where",
-           "nansum", "nanprod", "nanmin", "nanmax"]
